+++ cognite/experimental/_client.py
@@ -18,7 +18,11 @@ from cognite.experimental._api.pnid_parsing import PNIDParsingAPI
 from cognite.experimental._api.templatecompletion import ExperimentalTemplatesAPI
 
 APIClient._RETRYABLE_POST_ENDPOINT_REGEX_PATTERNS |= {
-    "^" + path + "(\?.*)?$" for path in ("/(types|labels|templates)/(list|byids|search)",)
+    "^" + path + "(\?.*)?$"
+    for path in (
+        "/(types|labels|templates)/(list|byids|search)",
+        "/alerts/deduplicate",
+    )
 }